+++ com.unity.cinemachine/CHANGELOG.md
@@ -32,6 +32,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - CinemachineIndependentImpulseListener renamed to CinemachineExternalImpulseListener.
 - Bugfix: Very occasional axis drift in SimpleFollow when viewing angle is +-90 degrees.
 - CinemachineExtension API changes to VirtualCamera, GetAllExtraStates, OnTargetObjectWarped, and ForceCameraPosition.
+- URP: add temporal effects reset on camera cut
+- Add Weight setting to CinemachinePostProcessing and CinemachineVolumeSettings
 
 
 ## [3.0.0-pre.3] - 2022-10-28